#2227 moved external link opening to ExternalLinkOpener class
