+++ core/src/main/java/org/kiwix/kiwixmobile/core/utils/Constants.kt
@@ -19,7 +19,6 @@ package org.kiwix.kiwixmobile.core.utils
 
 const val TAG_KIWIX = "kiwix"
 const val CONTACT_EMAIL_ADDRESS = "android@kiwix.org"
-const val KIWIX_SUPPORT_URL = "https://www.kiwix.org/support"
 
 // Request stuff
 const val REQUEST_STORAGE_PERMISSION = 1
